# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cloudflare Storage Service",
    "version": "1.0.0",
    "description": "Stores and serves files on Cloudflare R2"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
...
      "UploadResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "File record UUID"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Permanent public URL"
          },
          "size": {
            "type": "integer",
            "description": "File size in bytes"
          },
          "contentType": {
            "type": "string",
            "description": "MIME type"
          }
        },
        "required": [
          "id",
          "url",
          "size",
          "contentType"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "UploadRequest": {
        "type": "object",
        "properties": {
          "sourceUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL to download the file from"
          },
          "folder": {
            "type": "string",
            "description": "R2 key prefix/folder"
          },
          "filename": {
            "type": "string",
            "description": "Desired filename"
          },
          "contentType": {
            "type": "string",
            "description": "MIME type"
          }
        },
        "required": [
          "sourceUrl"
        ]
      },
      "UploadBase64Request": {
        "type": "object",
        "properties": {
          "contentBase64": {
            "type": "string",
            "description": "Base64-encoded file content. Data URL prefixes are accepted."
          },
          "folder": {
            "type": "string",
            "description": "R2 key prefix/folder"
          },
          "filename": {
            "type": "string",
            "description": "Desired filename"
          },
          "contentType": {
            "type": "string",
            "description": "MIME type"
          }
        },
        "required": [
          "contentBase64"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/upload": {
      "post": {
        "summary": "Upload a file to R2",
        "description": "Downloads a file from a given URL and uploads it to Cloudflare R2",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal org UUID"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal user UUID"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID from runs-service"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated list of brand UUIDs (e.g. uuid1,uuid2,uuid3)",
              "example": "00000000-0000-0000-0000-000000000001,00000000-0000-0000-0000-000000000002"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upload failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/upload/base64": {
      "post": {
        "summary": "Upload base64 content to R2",
        "description": "Decodes base64 file content from the request body and uploads it to Cloudflare R2",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal org UUID"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal user UUID"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID from runs-service"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated list of brand UUIDs (e.g. uuid1,uuid2,uuid3)",
              "example": "00000000-0000-0000-0000-000000000001,00000000-0000-0000-0000-000000000002"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadBase64Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upload failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}